Move the initCsvTable code from Lorre main function
