+++ src/main/scala/tech/cryptonomic/conseil/Lorre.scala
@@ -7,7 +7,6 @@ import akka.stream.ActorMaterializer
 import mouse.any._
 import com.typesafe.scalalogging.LazyLogging
 import org.slf4j.LoggerFactory
-import slick.lifted.{AbstractTable, TableQuery}
 import tech.cryptonomic.conseil.tezos.{
   ApiOperations,
   DatabaseConversions,
@@ -22,9 +21,8 @@ import tech.cryptonomic.conseil.tezos.{
 }
 import tech.cryptonomic.conseil.tezos.TezosTypes._
 import tech.cryptonomic.conseil.io.MainOutputs.LorreOutput
-import tech.cryptonomic.conseil.util.{ConfigUtil, DatabaseUtil}
+import tech.cryptonomic.conseil.util.DatabaseUtil
 import tech.cryptonomic.conseil.config._
-import tech.cryptonomic.conseil.tezos.TezosDatabaseOperations.insertWhenEmpty
 import tech.cryptonomic.conseil.tezos.TezosNodeOperator.{FetchRights, LazyPages}
 
 import scala.concurrent.duration._
@@ -89,7 +87,7 @@ object Lorre extends App with TezosErrors with LazyLogging with LorreAppConfig w
   implicit val tokenContracts: TokenContracts = {
 
     val futureTokenContracts =
-      initTableFromCsv(Tables.RegisteredTokens, tezosConf.network, separator = '|').map {
+      TezosDb.initTableFromCsv(db, Tables.RegisteredTokens, tezosConf.network, separator = '|').map {
         case (tokenRows, _) =>
           TokenContracts.fromTokens(
             tokenRows.map {
@@ -104,27 +102,7 @@ object Lorre extends App with TezosErrors with LazyLogging with LorreAppConfig w
   }
 
   /** Inits tables with values from CSV files */
-  initTableFromCsv(Tables.KnownAddresses, tezosConf.network)
-
-  import shapeless._
-  import shapeless.ops.hlist._
-
-  import kantan.csv._
-
-  /** Reads and inserts CSV file to the database for the given table */
-  def initTableFromCsv[A <: AbstractTable[_], H <: HList](table: TableQuery[A], network: String, separator: Char = ',')(
-      implicit hd: HeaderDecoder[A#TableElementType],
-      g: Generic.Aux[A#TableElementType, H],
-      m: Mapper.Aux[ConfigUtil.Csv.Trimmer.type, H, H]
-  ): Future[(List[A#TableElementType], Option[Int])] = {
-    val rows = ConfigUtil.Csv.readTableRowsFromCsv(table, network, separator)
-    db.run(insertWhenEmpty(table, rows))
-      .andThen {
-        case Success(_) => logger.info("Written {} {} rows", rows.size, table.baseTableRow.tableName)
-        case Failure(e) => logger.error(s"Could not fill ${table.baseTableRow.tableName} table", e)
-      }
-      .map(rows -> _)
-  }
+  TezosDb.initTableFromCsv(db, Tables.KnownAddresses, tezosConf.network)
 
   /** Schedules method for fetching baking rights */
   system.scheduler.schedule(lorreConf.blockRightsFetching.initDelay, lorreConf.blockRightsFetching.interval)(
